+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Changes
 
 - If an item can be looked up by Hey Wiki, a tooltip is added to the item in the inventory. This can be configured.
+- Added `/whatversion` command to open the Minecraft Wiki page for the current game version.